% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_folders.R
\name{list_folders}
\alias{list_folders}
\title{List Folders}
\usage{
list_folders(folder_path, one_kb_filter = T, file_names_filter = T)
}
\arguments{
\item{folder_path}{A character file path containing subfolders labeled
by family id which contain .pkl files.}

\item{one_kb_filter}{A boolean toggle set to TRUE by default. When toggled to TRUE
any .pkl files of size 1KB or less will be noted. These are potentially empty files.
The user should investigate these files, and if truly empty, delete them and
rerun this function before inputting into the pipeline.}

\item{file_names_filter}{A boolean toggle set to TRUE by default. When toggled to TRUE
any .pkl files that do not follow the standard naming scheme are noted.
The naming scheme requires all files to start with a numeric family code,
followed by _, followed by an alphanumeric person code.
Incorrectly named files may not be properly processed.
The user should investigate these files, and address naming concerns then
rerun this function before inputting into the pipeline.}
}
\value{
A character vector of file paths that contain .pkl files to be used as input into
the folder_path argument of data_pipeline.
}
\description{
Creates a list of all files containing .pkl files ready to be processed.
}
\examples{
list_folders('X:/Daily_2/ABC/tottag R code/Chris Processed/Batch Test/testData')

}
